Experimental: GNOME Shell 45+ local fixes
Provided by Codex (GPT-5.2) and tested on Ubuntu 25.10.
